SML_MessageBody tags are u32

+++ sml/include/sml/sml_message.h
@@ -62,7 +62,7 @@ extern "C" {
 #endif
 
 typedef struct {
-	u16 *tag;
+	u32 *tag;
 	void *data;
 } sml_message_body;
 
@@ -83,7 +83,7 @@ void sml_message_write(sml_message *msg, sml_buffer *buf);
 
 // SML_MESSAGE_BODY
 sml_message_body *sml_message_body_parse(sml_buffer *buf);
-sml_message_body *sml_message_body_init(u16 tag, void *data);
+sml_message_body *sml_message_body_init(u32 tag, void *data);
 void sml_message_body_free(sml_message_body *message_body);
 void sml_message_body_write(sml_message_body *message_body, sml_buffer *buf);
